requirements.txt
pandas==2.2.3
openpyxl==3.1.3
reportlab==4.0.0
cx_Freeze==7.2.7
zebra==0.2.0
Pillow==10.3.0
requests~=2.32.3
plyer~=2.1.0
python-dotenv~=1.0.1
PyYAML~=6.0.2
pystray~=0.19.5
fpdf~=1.7.2